+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,14 +1,14 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
 import static java.util.stream.Collectors.*;
 
 import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Product;
-import kitchenpos.dto.MenuProductSaveRequest;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.MenuSaveRequest;
+import kitchenpos.menu.dto.MenuProductSaveRequest;
+import kitchenpos.menu.dto.MenuResponse;
+import kitchenpos.menu.dto.MenuSaveRequest;
 import kitchenpos.repository.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
