+++ src/main/java/kitchenpos/menu/dto/MenuProductSaveRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 public class MenuProductSaveRequest {
