+++ src/main/java/kitchenpos/menu/repository/JpaMenuDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuDao extends JpaRepository<Menu, Long>, MenuDao {
