+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
-import kitchenpos.application.MenuService;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.MenuSaveRequest;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.dto.MenuResponse;
+import kitchenpos.menu.dto.MenuSaveRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
