+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,29 +1,27 @@
 package kitchenpos.application;
 
-import static kitchenpos.application.fixture.MenuFixtures.generateMenu;
-import static kitchenpos.application.fixture.MenuGroupFixtures.generateMenuGroup;
-import static kitchenpos.application.fixture.MenuProductFixtures.generateMenuProduct;
-import static kitchenpos.application.fixture.OrderFixtures.generateOrder;
-import static kitchenpos.application.fixture.OrderFixtures.generateOrderChangeOrderStatusRequest;
-import static kitchenpos.application.fixture.OrderFixtures.generateOrderSaveRequest;
-import static kitchenpos.application.fixture.OrderLineItemFixtures.*;
-import static kitchenpos.application.fixture.OrderTableFixtures.generateOrderTable;
-import static kitchenpos.application.fixture.ProductFixtures.generateProduct;
+import static kitchenpos.common.fixture.MenuFixtures.generateMenu;
+import static kitchenpos.common.fixture.MenuGroupFixtures.generateMenuGroup;
+import static kitchenpos.common.fixture.MenuProductFixtures.generateMenuProduct;
+import static kitchenpos.common.fixture.OrderFixtures.generateOrder;
+import static kitchenpos.common.fixture.OrderFixtures.generateOrderChangeOrderStatusRequest;
+import static kitchenpos.common.fixture.OrderFixtures.generateOrderSaveRequest;
+import static kitchenpos.common.fixture.OrderTableFixtures.generateOrderTable;
+import static kitchenpos.common.fixture.ProductFixtures.generateProduct;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.dao.MenuDao;
+import kitchenpos.menu.repository.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Product;
