+++ src/test/java/kitchenpos/common/fixture/MenuFixtures.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.fixture;
+package kitchenpos.common.fixture;
 
 import static java.util.stream.Collectors.*;
 
@@ -6,11 +6,11 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Price;
-import kitchenpos.dto.MenuProductSaveRequest;
-import kitchenpos.dto.MenuSaveRequest;
+import kitchenpos.menu.dto.MenuProductSaveRequest;
+import kitchenpos.menu.dto.MenuSaveRequest;
 
 public class MenuFixtures {
