+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -1,8 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
-import static kitchenpos.application.fixture.MenuGroupFixtures.generateMenuGroup;
-import static kitchenpos.application.fixture.MenuProductFixtures.generateMenuProduct;
-import static kitchenpos.application.fixture.ProductFixtures.*;
+import static kitchenpos.common.fixture.MenuGroupFixtures.generateMenuGroup;
+import static kitchenpos.common.fixture.MenuProductFixtures.generateMenuProduct;
+import static kitchenpos.common.fixture.ProductFixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -10,6 +10,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.common.UnitTest;
+import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.Product;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
